requirements.txt
flask
flask-cors
pyjwt
cryptography
python-dotenv